pyproject.toml
[project]
name = "vocab-trainer"
version = "0.1.0"
description = "Adaptive vocabulary learning system using Elo rating"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
